+++ src/taskTracker/tasks/Epic.java
@@ -1,30 +1,34 @@
-package TaskTracker.Tasks;
+package taskTracker.tasks;
 
+import java.util.HashMap;
 import java.util.Objects;
 
-public final class Subtask extends Task {
-    private Epic parentTask;
+public final class Epic extends Task {
+    private final HashMap<Integer, Subtask> subtasks;
 
-    public Subtask(String name, String description, TaskStatus status) {
+    public Epic(String name, String description, TaskStatus status,
+                HashMap<Integer, Subtask> subtasks) {
         super(name, description, status);
+        this.subtasks = subtasks;
     }
 
-    public Epic getParentTask() {
-        return parentTask;
+    public HashMap<Integer, Subtask> getSubtasks() {
+        return subtasks;
     }
-    public void setParentTask(Epic task) {
-        this.parentTask = task;
+
+    public void clearSubTasks() {
+        subtasks.clear();
     }
 
     @Override
     public boolean equals(Object object) {
         if (object == null) return false;
         if (this != object && getClass() != object.getClass()) return false;
-        Subtask otherTask = (Subtask) object;
+        Epic otherTask = (Epic) object;
         return Objects.equals(this.id, otherTask.getId())&&
                 Objects.equals(this.name, otherTask.name)&&
                 Objects.equals(this.description, otherTask.description) &&
-                Objects.equals(this.parentTask, otherTask.getParentTask());
+                Objects.equals(this.subtasks, otherTask.subtasks);
     }
 
     @Override
@@ -41,8 +45,8 @@ public final class Subtask extends Task {
         if(description != null) {
             hash += description.hashCode();
         }
-        if(parentTask != null){
-            hash += parentTask.hashCode();
+        if(subtasks != null){
+            hash += subtasks.hashCode();
         }
         return hash;
     }
